[SourceKitStressTester] Fix processAsync to fall back to sync processing when a failure is encountered
When any request takes down the service, all requests report failure, so it's
unclear which location actually triggered the crash. Crashes are uncommon
though, so to keep the runtime improvements, we still send requests
asynchronously but fall back to serially checking each offset when something
fails to determine which location was the cause.