Copy build-script-helper.py up to project root
I’m temporarily leaving the old SourceKitStressTester build-script-helper in place so I don’t have to coordinate breaking changes to apple/swift’s utils/build-script.

The copied script has not yet been modified to be usable in this location, but committing this version will make it easier to see what changed when reviewing this pull request.