[SourceKitStressTester] Fix ConcurrentRewriteActionGenerator producing a 0-length RangeInfo Action
It incorrectly assumed the start token of a range completed by the end token
would always be within the same token group (i.e. the same top-level
declaration). This is true for all tokens except the last one in the file,
which also completes the range of the entire file. Its start token is always
the first token of the first group/top-level declaration.