[SourceKitStressTester] Fix completion expected-result-checking false positive
For source code like "f(45)", at the position of 'f' we produce an expected
completion result with name "f(_:)". If 'f' is a function, code completion
produces a result with that name, and we find it. To handle the case where it's
a variable of function type, where the name is just "f" rather than "f(_:), we
were ignoring the argument labels if the completion result was a variable and
its type string contained "->". The type string can be a simple identifier
that's a typealias of a function type though, so we were incorrectly reporting
an issue in those cases.  This patch changes the logic to match on base name for
any variable result if the expected name argument labels are all "_".