Add request to test modules with new allow errors flag
A new request "TestModule" runs a compile with the Swift frontend flag
"-experimental-allow-module-with-compiler-errors". It then uses
sourcekitd to generate the interface for the compiled module, checking
that it can be read back in without errors.

The test module action is skipped for tokens within function bodies as
its assumed these will not be parsed/typechecked in the actual use case
anyway.