Treat string literals as monolithic for inside out rewrite
When rewriting the In the `APIService.swift` file from the `MovieSwift` source compatibility project, we currently end up with a source state that contains the following text, consisting mostly out of the string literals in the `switch` statement:
```
.movie/popularperson/popularmovie/top_ratedmovie/upcomingmovie/now_playingtrending/movie/day.movieDetail(movie)movie/\(String(movie)).videos(movie)movie/\(String(movie))/videos.personDetail(person)person/\(String(person)).credits(movie)movie/\(String(movie))/credits.review(movie)movie/\(String(movie))/reviews.recommended(movie)movie/\(String(movie))/recommendations.similar(movie)movie/\(String(movie))/similar.personMovieCredits(person)person/\(person)/movie_credits.personImages(person)person/\(person)/imagessearch/moviesearch/keywordsearch/persongenre/movie/listdiscover/movie
```

This expression takes about 5 minutes to typecheck for codecompletion after the leading dot (it will eventually) and we will eventually hit our self-imposed timeout.

IMHO there is little benefit in stress testing the statement above where `/` operators and variables are actually just string literal contents.

To avoid hitting this behaviour the issue, don’t increase the depth for constant parts of a string literal (i.e. non-interpolated parts). That way these constant parts will always be added together with their surrounding quotes.