Continue running the stress tester after an error es encountered
The stress tester runs stable enough on the test suite now that a failure isn’t likely to cause a whole bunch of following failures.

Instead, continue running after a failures is encountered to get a more complete picture of the SourceKit issues in the source compatibility suite. The current behaviour of uncovering new issues after XFailed issues are fixed is annoying.