Add a soft timeout if a request’s duration is close to the timeout
Requests that time out are flaky if their duration is close to the timeout level, because they will sometimes finish within the allowed time and sometimes fail. This is especially true if the machine is under heavy load, which is common when stress testing a project.

To mitigate some of this flakiness, introduce soft errors: Whenever a requests takes more than half of the allowed time, throw a soft error. This error will match an XFail, but will not fail the stress tester on its own.

So the idea is to have all of the potential timeout cases in the XFails file. We will only be required to remove it from there, if its execution duration has improved significantly (by more than 2x)