Don't filter results when opening a codecompletion session
Filtering the results in SourceKit takes considerable time, which we are not really interested in measuring.

Retrieve the initial set of results without a filter text with a result limit of 200. If we need to check for an expected result, perform a subsequent code completion update that sets the filter text. Don't measure the performance of that request.