Report SourceKit errors even if sk-stress-test returned exit code 0
I think this is what was causing the weird insideOut-893 failure: If the sk-stress-test invocation of insideOut-893 did not hit any other timeouts (only soft timeouts), it would return with exit code 0. Thus, sk-swiftc-wrapper would not report the soft timeout for insideOut-893 to the issue manager and its XFail would not be marked as matched.

So essentially, the insideOut-893 XFail was matched whenever any *other* request in the sk-stress-test invocation that also executes insideOut-893 properly timed out. This should explain the apparent flakiness.