Don’t add rpath to toolchain stdlib
We are only building the stress tester for macOS and since the stdlib is ABI stable we should be able to just use the system’s stdlib. This works around an issue where `install_name_tool` fails to add an rpath because `larger updated load commands do not fit`.

rdar://120675067